make dedupe_ctxt blocksize 64 bits
Thanks to github user juliantaylor for pointing this out -
set_aligned_same_length() uses this variable to mask out the lower bits of
an unaligned length. If the blocksize variable is 32 bits though and the
length is 64, we'll get the top bits of our blocksize mask as zeros, which
gives us wildly incorrect answers.

Git-commit: 0a9771f59daba95bae6ead2cafccdf0205279c88

+++ dedupe.h
@@ -25,7 +25,12 @@ struct dedupe_ctxt {
 
 	unsigned int		same_size;
 
-	unsigned int		fs_blocksize;
+	/*
+	 * Keep blocksize 64 bits wide here so we can use it for
+	 * aligning the (64 bits wide) context length.
+	 */
+	uint64_t		fs_blocksize;
+
 	/*
 	 * request tracking.
 	 *	queued: request is awaiting dedupe
